pyproject.toml
[project]
name = "syntts-asr"
version = "0.1.0"
description = "Synthetic Speech Augmentation for Low-Resource European ASR: Estonian and Slovenian"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "yuriyvnv"},
]

dependencies = [
    "openai>=1.75.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "datasets>=2.14.0",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
    "accelerate>=0.25.0",
    "jiwer>=3.0.0",
    "wandb>=0.16.0",
    "scipy>=1.11.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "soundfile>=0.12.0",
    "librosa>=0.10.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "torchcodec>=0.10.0",
]

[project.optional-dependencies]
nemo = [
    "nemo_toolkit[asr]>=2.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
